requirements.txt
python-dotenv~=1.1.1
huggingface-hub~=0.33.2
discord.py==2.5.2
requests==2.32.4
beautifulsoup4~=4.13.4